Moved daemon.py Share to separate file to avoid import cycle.
